feat(core): protect the route when the user is not connected
