+++ src/core/authentification/auth.hook.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+import { AuthContext } from './AuthProvider';
+
+export const useAuth = () => {
+  const currentUserContext = useContext(AuthContext);
+  return currentUserContext;
+};
