--- src/shared/components/Protected.tsx
@@ -1,13 +1,0 @@
-import { ReactNode } from 'react';
-import { Navigate } from 'react-router-dom';
-
-interface Props {
-  isSignedIn: boolean;
-  children: ReactNode;
-}
-export default function Protected({ isSignedIn, children }: Props) {
-  if (!isSignedIn) {
-    return <Navigate to='/' replace />;
-  }
-  return <>{children}</>;
-}
