+++ src/shared/hooks/local-storage.hook.ts
@@ -1,0 +1,29 @@
+import { useState } from 'react';
+
+export const useLocalStorage = <T>(keyName: string, defaultValue: T) => {
+  const [storedValue, setStoredValue] = useState(() => {
+    try {
+      const value = window.localStorage.getItem(keyName);
+
+      if (value) {
+        return JSON.parse(value);
+      } else {
+        window.localStorage.setItem(keyName, JSON.stringify(defaultValue));
+        return defaultValue;
+      }
+    } catch (err) {
+      return defaultValue;
+    }
+  });
+
+  const setValue = <T>(newValue: T) => {
+    try {
+      window.localStorage.setItem(keyName, JSON.stringify(newValue));
+    } catch (err) {
+      undefined;
+    }
+    setStoredValue(newValue);
+  };
+
+  return [storedValue, setValue];
+};
